newlib-1.18.0-rtems4.10-20110518.diff

+++ newlib/libc/string/memmove.c
@@ -36,13 +36,14 @@ QUICKREF
 */
 
 #include <string.h>
+#include <stdint.h>
 #include <_ansi.h>
 #include <stddef.h>
 #include <limits.h>
 
 /* Nonzero if either X or Y is not aligned on a "long" boundary.  */
 #define UNALIGNED(X, Y) \
-  (((long)X & (sizeof (long) - 1)) | ((long)Y & (sizeof (long) - 1)))
+  (((intptr_t)X & (sizeof (long) - 1)) | ((intptr_t)Y & (sizeof (long) - 1)))
 
 /* How many bytes are copied each iteration of the 4X unrolled loop.  */
 #define BIGBLOCKSIZE    (sizeof (long) << 2)
